Refactor: Move buildBundle to the LuviAppBundle module
